# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Sankara Eye Hospital Vision 2020 Conference 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /system-users:
    get:
      operationId: listSystemUsers
      tags: [auth]
      summary: List system users (admin, coordinators, committee)
      responses:
        "200":
          description: System users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SystemUser"
    post:
      operationId: createSystemUser
      tags: [auth]
      summary: Create a system user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemUserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemUser"
  /system-users/{id}:
    patch:
      operationId: updateSystemUser
      tags: [auth]
      summary: Update a system user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemUserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemUser"
    delete:
      operationId: deleteSystemUser
      tags: [auth]
      summary: Delete a system user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    SystemUser:
      type: object
      required: [id, empId, name, userType, mustChangePassword, createdAt]
      properties:
        id:
          type: integer
        empId:
          type: string
        name:
          type: string
        email:
          type: ["string", "null"]
        mobile:
          type: ["string", "null"]
        userType:
          type: string
          enum: [admin, super_admin, track_coordinator, food_coordinator, scientific_committee, pr_member, coordinator_view_only]
        assignedTrack:
          type: ["string", "null"]
        mustChangePassword:
          type: boolean
        createdAt:
          type: string
          format: date-time
    SystemUserInput:
      type: object
      required: [empId, name, userType]
      properties:
        empId:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        userType:
          type: string
          enum: [admin, super_admin, track_coordinator, food_coordinator, scientific_committee, pr_member, coordinator_view_only]
        password:
          type: string
          description: Defaults to Welcome@123 if omitted
        assignedTrack:
          type: string
    SystemUserUpdate:
      type: object
      properties:
        empId:
          type: string
        name:
          type: string
        email:
          type: string
        mobile:
          type: string
        userType:
          type: string
          enum: [admin, super_admin, track_coordinator, food_coordinator, scientific_committee, pr_member, coordinator_view_only]
        assignedTrack:
          type: string
        password:
          type: string